# Fix benchmark output when scheduler address is specified