--- a/dask_cuda/benchmarks/local_cupy_transpose_sum.py
+++ b/dask_cuda/benchmarks/local_cupy_transpose_sum.py
@@ -115,7 +115,7 @@ async def run(args):
             for (d1, d2), bw in sorted(bandwidths.items()):
                 fmt = (
                     "(%s,%s)     | %s %s %s (%s)"
-                    if args.multi_node
+                    if args.multi_node or args.sched_addr
                     else "(%02d,%02d)     | %s %s %s (%s)"
                 )
                 print(fmt % (d1, d2, bw[0], bw[1], bw[2], total_nbytes[(d1, d2)]))