With this PR we fix a failure in the output of the benchmark:

Before:

```bash
(w1,w2)     | 25% 50% 75% (total nbytes)
-------------------------------
Traceback (most recent call last):
  File "local_cudf_merge.py", line 349, in <module>
    main(parse_args())
  File "local_cudf_merge.py", line 280, in main
    print(fmt % (d1, d2, bw[0], bw[1], bw[2], total_nbytes[(d1, d2)]))
TypeError: %d format: a number is required, not str
```

After:

```bash
(20200918) quasiben@DirtyHorseZ:~/Github/dask-cuda/dask_cuda/benchmarks$ python local_cudf_merge.py -c 100000 --scheduler-address 192.168.1.41:8786
Merge benchmark
-------------------------------
backend        | dask
merge type     | gpu
rows-per-chunk | 100000
protocol       | tcp
device(s)      | 0
rmm-pool       | True
frac-match     | 0.3
data-processed | 6.40 MB
===============================
Wall-clock     | Throughput
-------------------------------
79.10 ms       | 80.91 MB/s
76.30 ms       | 83.88 MB/s
75.31 ms       | 84.99 MB/s
===============================
(w1,w2)     | 25% 50% 75% (total nbytes)
-------------------------------
(tcp://192.168.1.41:35669,tcp://192.168.1.41:36401)     | 200.85 MB/s 251.42 MB/s 276.78 MB/s (72.00 MB)
(tcp://192.168.1.41:36401,tcp://192.168.1.41:35669)     | 177.83 MB/s 232.73 MB/s 285.58 MB/s (72.00 MB)
```