@pentschev 2020-10-14
I just realized we had to fix it for CuPy benchmarks too, I pushed the fix to your branch @quasiben , hope you don't mind.